Add ImageUpload components with ImagePreview, ImageRemoveButton, and ImageUploadHandler for enhanced image handling functionality
