+++ src/components/inputs/ImageUpload/ImageRemoveButton.tsx
@@ -1,0 +1,14 @@
+import { Button } from "@mantine/core";
+import React from "react";
+
+type Props = {
+    removeImage: () => void;
+};
+
+export default function ImageRemoveButton(props: Props) {
+    return (
+        <Button type="button" variant="light" size="xs" onClick={props.removeImage}>
+            Remove
+        </Button>
+    );
+}
