+++ src/components/inputs/ImageUpload/ImageUploadButton.tsx
@@ -1,0 +1,14 @@
+import { Button } from "@mantine/core";
+import React from "react";
+
+type Props = {
+    onClick: () => void;
+};
+
+export default function ImageUploadButton(props: Props) {
+    return (
+        <Button type="button" variant="light" size="xs" onClick={props.onClick}>
+            Upload
+        </Button>
+    );
+}
